+++ pymilo/transporters/labelbinarizer_transporter.py
@@ -77,7 +77,8 @@ class LabelBinarizerTransporter(AbstractTransporter):
         :return: a sklearn.preprocessing.LabelBinarizer instance derived from the
         pymilo deserialized output of the previously pymilo serialized label_binarizer field.
         """
-        raw_lb = KEYS_NEED_PREPROCESSING_BEFORE_DESERIALIZATION["_label_binarizer"]()
+        raw_lb = KEYS_NEED_PREPROCESSING_BEFORE_DESERIALIZATION["_label_binarizer"](
+        )
         for item in content.keys():
             setattr(raw_lb, item, content[item])
         return raw_lb