+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- GeneralDataStructure transporter enhanced
+  - Deep ndarray to list serialization
+  - Deep list to ndarray deserialization 
+- LabelBinerizer transporter updated
 ## [0.1] - 2023-06-29
 ### Added
 - scikit-learn linear models support