+++ CHANGELOG.md
@@ -5,9 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `ndarray_to_list` method added to `GeneralDataStructureTransporter`
+- `list_to_ndarray` method added to `GeneralDataStructureTransporter` 
+### Changed
 - GeneralDataStructure transporter enhanced
-  - Deep ndarray to list serialization
-  - Deep list to ndarray deserialization 
 - LabelBinerizer transporter updated
 ## [0.1] - 2023-06-29
 ### Added
@@ -18,3 +20,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased]: https://github.com/openscilab/pymilo/compare/v0.1...dev
 [0.1]: https://github.com/openscilab/pymilo/compare/e887108...v0.1
+