+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `ndarray_to_list` method added to `GeneralDataStructureTransporter`
 - `list_to_ndarray` method added to `GeneralDataStructureTransporter` 
-- NN chain initialized
+- `neural_network_chain.py` chain
 ### Changed
 - GeneralDataStructure transporter enhanced
 - LabelBinerizer transporter updated