+++ CHANGELOG.md
@@ -8,9 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `ndarray_to_list` method added to `GeneralDataStructureTransporter`
 - `list_to_ndarray` method added to `GeneralDataStructureTransporter` 
+- `neural_network_chain.py` chain
 ### Changed
 - GeneralDataStructure transporter enhanced
 - LabelBinerizer transporter updated
+- transporters' chain router added to `pymilo func`
+- NeuralNetwork params initialized in `pymilo_param`
+- `pymilo_test` updated to support multiple models
+- `linear_model_chain` refactored
 ## [0.1] - 2023-06-29
 ### Added
 - scikit-learn linear models support