+++ CHANGELOG.md
@@ -42,6 +42,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased]: https://github.com/openscilab/pymilo/compare/v0.2...dev
 [0.1]: https://github.com/openscilab/pymilo/compare/e887108...v0.1
-[0.2]: https://github.com/openscilab/pymilo/compare/v.0.1...v0.2
+[0.2]: https://github.com/openscilab/pymilo/compare/v0.1...v0.2